requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4